autocalibration/pyproject.toml
[project]
name = "autocalibration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "tqdm>=4.0",
    "requests>=2.32.5",
    "blenderproc>=2.8.0",
]
